Set User ID add `find` parsing
Previously `_check_session()` was only being called when provided with
a session ID. This was because the point of that method was to hit the
session endpoint, while cookie-based login was already getting a
response from T1. So we can make them both parse the user ID and
session ID.

Additionally, `find` wasn’t working when passed in `True` or `False`,
because `str(True) == “True”`, but T1 is expecting 1 or 0.

Last, fix some whitespace issues.