Remove unknown items before POST
Part of the API breaking changes is that addition of a new parameter
does not constitute a breaking change. This is fine, we don’t care on
read. But, previously we would post unknown items on save—we just
wouldn’t do any validation.

But this breaks for new parameters that are read-only on update; with
these, we would send them along anyway and get an error. Now, if we
don’t know what an attribute is beforehand, we ignore it.