Set dummy version in TargetDimension.save
TargetDimension doesn’t have a version associated. This is expected, as
they don’t have them from the E&M API. However, Entity.save expects
there to be a version number attached. Setting it to None will suffice
to get the version, and then _validate_write will throw it out (because
it’s not in _push).

This strategy can be used for all entities that should not have a
version included.