Child pixel removal and strategy start-/end-date fix
- Strategy start/end dates have a weird dual setup. You can choose to
use the campaign dates or set custom ones. If the former, you cannot
post custom start-end dates, and so `Strategy.save` removes those from
the body. However, if you are changing a strategy from having start/end
dates to using campaign start/end dates, you are supposed to POST a
null value. So we need to convert the None to an empty string.