Add property check to enable pre-addition validation.
Sometimes (for instance, if you're receiving user input), you need to make sure
that properties are valid properties for an entity to have. Rather than using
the _pull dict (which is the canonical place all properties exist) directly,
wrap it in a method (so that if we change it later, we know this is a
dependency).